requirements.txt
Django==1.10.6
Jinja2==2.9.6
MarkupSafe==1.0
PyJWT==1.5.3
coreapi==2.3.1
coreschema==0.0.4
defusedxml==0.5.0
dj-static==0.0.6
django-allauth==0.33.0
# django-rest-auth==0.9.1
django-rest-swagger==2.1.2
djangocms-admin-style==1.2.7
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
elasticsearch==5.4.0
gunicorn==19.7.1
itypes==1.1.0
jieba==0.39
oauthlib==2.0.4
openapi-codec==1.3.2
pandas==0.21.0
psycopg2==2.7.1
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
simplejson==3.11.1
six==1.11.0
static3==0.7.0
tensorflow==1.4.0
uritemplate==3.0.0
asgi-redis==1.4.3
channels==1.1.8
numpy==1.13.3
pytz==2017.3
bs4==0.0.1
lxml==4.1.1
social-auth-app-django==2.0.0
